+++ tests/test_trie.py
@@ -11,11 +11,15 @@ def test_insert_find() -> None:
 
     text = "set to 10"
     results = list(trie.find(text))
-    assert results == [(9, "10", 10)]
+    assert results == [(8, "1", 1), (9, "10", 10)]
     for end_pos, number_text, number_value in results:
         start_pos = end_pos - len(number_text)
         assert text[start_pos:end_pos] == number_text
         assert int(number_text) == number_value
+
+    # With word boundaries
+    results = list(trie.find(text, word_boundaries=True))
+    assert results == [(9, "10", 10)]
 
     assert list(trie.find("set to 1, then *two*, then finally twenty two please!")) == [
         (8, "1", 1),
@@ -43,11 +47,14 @@ def test_insert_find() -> None:
     assert not list(trie.find(""))
 
     # Test word boundaries
-    assert not list(trie.find("1two"))
-    assert list(trie.find("1,two")) == [(1, "1", 1), (5, "two", 2)]
+    assert not list(trie.find("1two", word_boundaries=True))
+    assert list(trie.find("1,two", word_boundaries=True)) == [
+        (1, "1", 1),
+        (5, "two", 2),
+    ]
 
     # ° is not a word boundary
-    assert not list(trie.find("10°"))
+    assert not list(trie.find("10°", word_boundaries=True))
 
 
 def test_multiple_values() -> None: